Refactor imports for @hedvig-ui/Card
