+++ shared/hedvig-ui/index.ts
@@ -7,6 +7,13 @@ import {
   ButtonsGroup,
   buttonSizeMap,
 } from './Button/button'
+import {
+  Card,
+  CardContent,
+  CardLink,
+  CardsWrapper,
+  DangerCard,
+} from './Card/card'
 import { Checkbox } from './Checkbox/checkbox'
 
 export {
@@ -18,4 +25,9 @@ export {
   ButtonLink,
   ButtonProps,
   Checkbox,
+  Card,
+  CardContent,
+  CardLink,
+  CardsWrapper,
+  DangerCard,
 }
