+++ src/components/claims/claim-details/index.tsx
@@ -1,9 +1,9 @@
 import styled from '@emotion/styled'
+import { Card, CardsWrapper } from '@hedvig-ui'
 import { ClaimState, useClaimPageQuery } from 'api/generated/graphql'
 import { ClaimItems } from 'components/claims/claim-details/components/claim-items'
 import { ChatPane } from 'components/member/tabs/ChatPane'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import { Card, CardsWrapper } from 'hedvig-ui/card'
 import { MainHeadline } from 'hedvig-ui/typography'
 import React, { useContext, useEffect } from 'react'
 import { Prompt, RouteComponentProps } from 'react-router'
