+++ src/components/member/tabs/account-tab/index.tsx
@@ -1,3 +1,4 @@
+import { Card, CardsWrapper } from '@hedvig-ui'
 import {
   ExpansionPanel,
   ExpansionPanelDetails,
@@ -16,7 +17,6 @@ import {
   LoadingMessage,
   StandaloneMessage,
 } from 'hedvig-ui/animations/standalone-message'
-import { Card, CardsWrapper } from 'hedvig-ui/card'
 import { InfoContainer, InfoRow, InfoText } from 'hedvig-ui/info-row'
 import { Spacing } from 'hedvig-ui/spacing'
 import { MainHeadline, ThirdLevelHeadline } from 'hedvig-ui/typography'
