+++ src/components/member/tabs/campaigns-tab/referrals/ReferralsInfo.tsx
@@ -1,3 +1,4 @@
+import { Card, CardsWrapper } from '@hedvig-ui'
 import { MemberReferral, ReferralInformation } from 'api/generated/graphql'
 import {
   BadgeRow,
@@ -5,7 +6,6 @@ import {
   TableHeadline,
 } from 'components/member/tabs/campaigns-tab/styles'
 import { Badge } from 'hedvig-ui/Badge/badge'
-import { Card, CardsWrapper } from 'hedvig-ui/card'
 import { Capitalized, Placeholder } from 'hedvig-ui/typography'
 import React from 'react'
 import { Link } from 'react-router-dom'
