+++ src/components/member/tabs/contracts-tab/contract/index.tsx
@@ -1,10 +1,10 @@
 import styled from '@emotion/styled'
+import { Card, CardsWrapper } from '@hedvig-ui'
 import { Contract as ContractType } from 'api/generated/graphql'
 import { Agreement } from 'components/member/tabs/contracts-tab/agreement'
 import { AgreementsTable } from 'components/member/tabs/contracts-tab/agreement/AgreementsTable'
 import { MasterInception } from 'components/member/tabs/contracts-tab/contract/master-inception'
 import { TerminationDate } from 'components/member/tabs/contracts-tab/contract/termination-date'
-import { Card, CardsWrapper } from 'hedvig-ui/card'
 import { InfoContainer, InfoRow, InfoText } from 'hedvig-ui/info-row'
 import { ThirdLevelHeadline } from 'hedvig-ui/typography'
 import React from 'react'
