+++ src/components/member/tabs/quote-tab/quote-sub-section.tsx
@@ -1,9 +1,8 @@
 import styled from '@emotion/styled'
-import { Button } from '@hedvig-ui'
+import { Button, Card, CardsWrapper } from '@hedvig-ui'
 import { Quote } from 'api/generated/graphql'
 import { CreateQuoteForm } from 'components/member/tabs/quote-tab/create-quote-form'
 import { useContracts } from 'graphql/use-contracts'
-import { Card, CardsWrapper } from 'hedvig-ui/card'
 import { MainHeadline } from 'hedvig-ui/typography'
 import React from 'react'
 import { ContractType, TypeOfContractType } from 'types/enums'
