+++ src/features/tools/campaign-codes/components/CampaignCodeTable.tsx
@@ -1,4 +1,5 @@
 import styled from '@emotion/styled'
+import { Card } from '@hedvig-ui'
 import {
   CampaignFilter,
   useSetCampaignCodeTypeMutation,
@@ -14,7 +15,6 @@ import {
   LoadingMessage,
   StandaloneMessage,
 } from 'hedvig-ui/animations/standalone-message'
-import { Card } from 'hedvig-ui/card'
 import { Popover } from 'hedvig-ui/popover'
 import { SearchableDropdown } from 'hedvig-ui/searchable-dropdown'
 import React from 'react'
