+++ src/features/tools/employees/index.tsx
@@ -1,8 +1,8 @@
 import styled from '@emotion/styled'
+import { Card, CardsWrapper } from '@hedvig-ui'
 import { useGetMeQuery } from 'api/generated/graphql'
 import { EmployeeFilter } from 'features/tools/employees/components/EmployeeFilter'
 import { EmployeeTable } from 'features/tools/employees/components/EmployeeTable'
-import { Card, CardsWrapper } from 'hedvig-ui/card'
 import { MainHeadline } from 'hedvig-ui/typography'
 import React, { useState } from 'react'
