+++ src/features/tools/index.tsx
@@ -1,10 +1,10 @@
 import styled from '@emotion/styled'
+import { CardLink, CardsWrapper } from '@hedvig-ui'
 import {
   StagingTools,
   stagingToolsAvailable,
 } from 'features/tools/staging-tools'
 import { FadeIn } from 'hedvig-ui/animations/fade-in'
-import { CardLink, CardsWrapper } from 'hedvig-ui/card'
 import React from 'react'
 
 const Icon = styled('div')`
